cloud_functions/gather/requirements.txt
google-cloud-firestore==1.6.1
